Catch naive date-time transformation to timezone-aware date failures
A naive timestamp issued during a daylight saving time (DST) transition
causes an internal server error.

The server is configured to use the Europe/Amsterdam timezone, and as
the timestamps are not explicitly zoned, that default timezone is used.
We can then test the behaviour of the server based on DST shifts that
happened in Amsterdam.

Change-Id: I59fd93c21e5fd87870a68c1c9c98b006ab79cdb2