Add management command to reset and update Hiccup statistics
Introduce a new management command `stats` that allows to reset and
update the Hiccup statistics easily. Resetting the statistics means
removing all of them and computing them again. Updating the statistics
means resuming counting from the last timestamp. To that end, a new
model `StatsMetadata` is created to hold the last known timestamp the
statistics were computed at.

The `stats` command is meant to be run at regular intervals. For example
with `cron`:

    0 0,6,12,18 * * * python /path/to/hiccup/manage.py stats update

FIXME:
    The counting algorithm drops duplicates based on their timestamp to
    avoid duplicate heartbeats. The true deduplication should happen on
    (`device_id`, `date`) instead. Better yet, the database should be
    cleared of those duplicates altogether.

Issue: HIC-112
Change-Id: I975c896bddbc1308d156f0dbd90d0b66d2d5529b